# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "To-Do List API",
    "version": "1.0.0",
...
    "contact": {
      "name": "API Support",
      "email": "support@todolist-api.com",
      "url": "https://github.com/your-org/todolist-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.todolist.com/api",
      "description": "Production server"
    }
  ],
...
  "components": {
    "schemas": {
      "List": {
        "type": "object",
        "required": ["id", "name", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the list (UUID v4)",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Name of the list (1-100 characters)",
            "example": "Work Tasks"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "Optional description of the list (max 500 characters)",
            "example": "Tasks related to work and professional development",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the list was created",
            "example": "2025-07-04T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the list was last updated",
            "example": "2025-07-04T10:30:00.000Z"
          },
          "taskCount": {
            "type": "integer",
            "minimum": 0,
            "description": "Computed field: number of tasks in this list",
            "example": 5,
            "nullable": true
          }
        }
      },
...
      "CreateListRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Name of the list (1-100 characters)",
            "example": "Personal Tasks"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "Optional description of the list (max 500 characters)",
            "example": "Personal tasks and household chores",
            "nullable": true
          }
        }
      },
      "UpdateListRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Name of the list (1-100 characters)",
            "example": "Updated Personal Tasks"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "description": "Optional description of the list (max 500 characters)",
            "example": "Updated description for personal tasks",
            "nullable": true
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "timestamp"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code for programmatic handling",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The provided data is invalid"
          },
          "details": {
            "description": "Additional error details (validation errors, etc.)",
            "example": {
              "field": "name",
              "constraint": "minLength",
              "value": ""
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Request timestamp when error occurred",
            "example": "2025-07-04T10:00:00.000Z"
          }
        }
      },
...
    },
...
  },
  "paths": {
    "/api/lists": {
      "get": {
        "summary": "Get all lists",
        "description": "Retrieves all lists with optional task counts",
        "tags": ["Lists"],
        "parameters": [
          {
            "in": "query",
            "name": "includeTaskCount",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Whether to include task counts for each list"
          }
        ],
        "responses": {
          "200": {
            "description": "List of all lists retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/List"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "example": 3
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new list",
        "description": "Creates a new list with the provided data",
        "tags": ["Lists"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateListRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "List created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/List"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or duplicate name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity - validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lists/{id}": {
      "get": {
        "summary": "Get a specific list by ID",
        "description": "Retrieves a specific list by its unique identifier",
        "tags": ["Lists"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The list ID"
          },
          {
            "in": "query",
            "name": "includeTaskCount",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Whether to include task count for the list"
          }
        ],
        "responses": {
          "200": {
            "description": "List retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/List"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid UUID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific list",
        "description": "Updates an existing list with the provided data",
        "tags": ["Lists"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The list ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/List"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error or duplicate name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity - validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific list",
        "description": "Deletes a list and all its associated tasks (cascade delete)",
        "tags": ["Lists"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The list ID"
          }
        ],
        "responses": {
          "204": {
            "description": "List deleted successfully"
          },
          "400": {
            "description": "Bad request - invalid UUID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}